pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "packaging"]
build-backend = "setuptools.build_meta"

[project]
name = "legacy-to-techdocs"
authors = [{ name = "Lena Garber", email = "lgarber@akamai.com" }]
description = "Quickly convert legacy Linode API docs URLs to TechDocs."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "Apache-2.0" }
classifiers = []
dependencies = [
    "openapi3",
    "PyYAML",
    "rich",
    "urllib3<3",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "build",
]

[project.scripts]
legacy-to-techdocs = "legacy_to_techdocs:main"

[tool.setuptools.packages.find]
include = ["legacy_to_techdocs*"]

[tool.isort]
profile = "black"
line_length = 80

[tool.black]
line-length = 80
target-version = ["py37", "py38", "py39", "py310", "py311"]
